Fjernet toggles for personlinje + oversiktsboks + endreBehandlingssattus + sendBrevEgetVindu
