+++ src/felleskomponenter/sideDialog/sideDialog.tsx
@@ -5,8 +5,6 @@ import { DokumentOversikt, FysiskDokument } from "Domene";
 
 import * as Nav from "../../navFrontend";
 import * as Utils from "../../utils";
-
-import { useFeatureToggle, Status } from "../../featuretoggle";
 import SideDialogSendBrev from "./sendBrev";
 import SideDialogOpprettNyBuc from "./sideDialogOpprettNyBuc";
 import SideDialogDokumenter from "./sideDialogDokumenter";
@@ -34,19 +32,11 @@ export const FaneViser = ({
   dokumentOversikt,
   dokumenter,
 }: FaneViserProps) => {
-  const sendBrevIEgetVinduToggle = useFeatureToggle("melosys.sendBrevIEgetVindu");
-
   switch (navn) {
     case "dokumenter":
       return <SideDialogDokumenter dokumentOversikt={dokumentOversikt} />;
     case "brevbestilling":
-      return (
-        <SideDialogSendBrev
-          behandlingID={behandlingID}
-          redigerbart={redigerbart}
-          visApneINyttVindu={sendBrevIEgetVinduToggle === Status.enabled}
-        />
-      );
+      return <SideDialogSendBrev behandlingID={behandlingID} redigerbart={redigerbart} />;
     case "sedbestilling":
       return <SideDialogOpprettNyBuc behandlingID={behandlingID} dokumenter={dokumenter} />;
     case "besvarsed":
