+++ src/sider/ftrl/saksbehandling/saksbehandling.js
@@ -14,11 +14,9 @@ import Personlinje from "../../../felleskomponenter/personlinje";
 import SideDialog from "../../../felleskomponenter/sideDialog/sideDialog";
 import SideOppsummering from "../../../felleskomponenter/oppsummering/sideOppsummering";
 import SaksoversiktLenke from "../../../felleskomponenter/saksoversiktLenke";
-import Behandlingsstatus from "../../../felleskomponenter/behandlingsstatus";
-import Legacybehandlingsmeny from "./legacybehandlingsmeny";
 import Stegvelger, { STEG } from "../../../felleskomponenter/stegvelger";
 import { SoknadMenypanelForm } from "../../../felleskomponenter/menypanelForm";
-import { useFeatureToggle, FeatureToggle } from "../../../featuretoggle";
+import { useFeatureToggle } from "../../../featuretoggle";
 
 import { fagsakOperations, fagsakSelectors } from "../../../ducks/fagsaker";
 import { behandlingsresultatOperations, behandlingsresultatSelectors } from "../../../ducks/behandlingsresultat";
@@ -40,61 +38,8 @@ import { AvslaattSoknad, HenlagtSak } from "../../eu_eøs/saksbehandling/kompone
 import { stegMap } from "./stegMap";
 import "./saksbehandling.css";
 
-const behandlingsstatusMap = {
-  [MKV.Koder.behandlinger.behandlingsstatus.VURDER_DOKUMENT]: [
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-      term: MKV.Terms.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-    },
-  ],
-  [MKV.Koder.behandlinger.behandlingsstatus.SVAR_ANMODNING_MOTTATT]: [
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-      term: MKV.Terms.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-    },
-  ],
-  [MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_UTL]: [
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-      term: MKV.Terms.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-    },
-  ],
-  [MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_PART]: [
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-      term: MKV.Terms.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-    },
-  ],
-  [MKV.Koder.behandlinger.behandlingsstatus.UNDER_BEHANDLING]: [],
-};
-
 const Saksbehandling = ({
   annenBehandlingOppfriskes,
-  apneTidligereBehandlinger,
   arbeidsland,
   behandlingOppfriskes,
   behandlingsgrunnlag,
@@ -102,8 +47,6 @@ const Saksbehandling = ({
   behandlingsgrunnlagPeriodeTom,
   behandlingsgrunnlagMottaksdato,
   behandlingsresultat,
-  behandlingstema,
-  behandlingsstatus,
   dokumenter,
   dokumentOversikt,
   fagsak,
@@ -120,14 +63,12 @@ const Saksbehandling = ({
   lagreAllData,
   lagreAvklartefakta,
   lagreBehandlingsgrunnlagOgOppfriskSaksopplysninger,
-  lagreOgLukk,
   lagreVilkar,
   landkoder,
   location,
   match,
   oppdaterBehandlingsgrunnlag,
   oppsummering,
-  person,
   redigerbart,
   resetBehandlingerState,
   resetBehandlingsgrunnlagState,
@@ -137,14 +78,9 @@ const Saksbehandling = ({
   resetMedlemskapsperiodeState,
   skjulMenypanel,
   soknadForm,
-  tilbakeleggOppgave,
   startOgVisOppfriskModal,
   tilForsiden,
-  visAvsluttSakSomBortfaltDialogHandle,
-  visAvslagSoknadDialogHandle,
-  visHenleggDialogHandle,
   visOppfriskModal,
-  visRevurderFagsakDialogHandle,
 }) => {
   const [behandlingID, setBehandlingID] = useState(-1);
   const [bestemmelser, setBestemmelser] = useState([]);
@@ -235,9 +171,7 @@ const Saksbehandling = ({
 
   return (
     <>
-      <FeatureToggle togglename="melosys.design.PERSONLINJE">
-        {(status) => status === "enabled" && <Personlinje />}
-      </FeatureToggle>
+      <Personlinje />
       <div id="main-container" className="main-container">
         <div className="saksbehandling">
           <Nav.Container fluid>
@@ -269,41 +203,16 @@ const Saksbehandling = ({
               </Nav.Column>
               <Nav.Column xs="5">
                 <SideOppsummering
-                  behandlingstema={behandlingstema}
-                  redigerbart={redigerbart}
-                  fagsak={fagsak}
                   oppsummering={oppsummering}
-                  person={person}
+                  fagsak={fagsak}
                   arbeidsland={
                     landkoder && landkoder.filter((landkodeObjekt) => arbeidsland.includes(landkodeObjekt.kode))
                   }
+                  lovvalgsperiodeFom={behandlingsgrunnlagPeriodeFom}
+                  lovvalgsperiodeTom={behandlingsgrunnlagPeriodeTom}
                   behandlingsgrunnlagPeriodeFom={behandlingsgrunnlagPeriodeFom}
                   behandlingsgrunnlagPeriodeTom={behandlingsgrunnlagPeriodeTom}
                   behandlingsgrunnlagMottaksdato={behandlingsgrunnlagMottaksdato}
-                  lovvalgsperiodeFom={behandlingsgrunnlagPeriodeFom}
-                  lovvalgsperiodeTom={behandlingsgrunnlagPeriodeTom}
-                  renderBehandlingsmeny={() => (
-                    <Legacybehandlingsmeny
-                      redigerbart={redigerbart}
-                      lagreOgLukkHandle={lagreOgLukk}
-                      tilbakeleggeHandle={tilbakeleggOppgave}
-                      oppfriskSaksopplysningerHandle={visOppfriskModal}
-                      visHenleggDialogHandle={visHenleggDialogHandle}
-                      visAvsluttSakSomBortfaltDialogHandle={visAvsluttSakSomBortfaltDialogHandle}
-                      visAvslagSoknadDialogHandle={visAvslagSoknadDialogHandle}
-                      apneTidligereBehandlinger={apneTidligereBehandlinger}
-                      visRevurderFagsakDialogHandle={visRevurderFagsakDialogHandle}
-                      behandlingsstatus={behandlingsstatus}
-                    />
-                  )}
-                  renderBehandlingsstatus={() => (
-                    <Behandlingsstatus
-                      behandlingID={behandlingID}
-                      redigerbart={redigerbart}
-                      oppsummering={oppsummering}
-                      behandlingsstatusMap={behandlingsstatusMap}
-                    />
-                  )}
                 />
                 <SaksoversiktLenke />
                 <SideDialog
@@ -331,8 +240,6 @@ Saksbehandling.propTypes = {
   behandlingsgrunnlagPeriodeTom: PT.string.isRequired,
   behandlingsgrunnlagMottaksdato: PT.string.isRequired,
   behandlingsresultat: MPT.Behandlingsresultat.isRequired,
-  behandlingstema: PT.string.isRequired,
-  behandlingsstatus: PT.string.isRequired,
   dokumenter: PT.array.isRequired,
   dokumentOversikt: PT.array.isRequired,
   fagsak: MPT.Fagsak,
@@ -342,11 +249,9 @@ Saksbehandling.propTypes = {
   location: PT.object.isRequired,
   match: PT.object.isRequired,
   oppsummering: MPT.Behandlinger.Oppsummering,
-  person: MPT.Behandlinger.Saksopplysninger.Person.isRequired,
   redigerbart: PT.bool,
   soknadForm: PT.object.isRequired,
   // Funcs
-  apneTidligereBehandlinger: PT.func.isRequired,
   hentBehandling: PT.func.isRequired,
   hentBehandlingsgrunnlag: PT.func.isRequired,
   hentBehandlingsresultat: PT.func.isRequired,
@@ -359,7 +264,6 @@ Saksbehandling.propTypes = {
   lagreAllData: PT.func.isRequired,
   lagreAvklartefakta: PT.func.isRequired,
   lagreBehandlingsgrunnlagOgOppfriskSaksopplysninger: PT.func.isRequired,
-  lagreOgLukk: PT.func.isRequired,
   lagreVilkar: PT.func.isRequired,
   oppdaterBehandlingsgrunnlag: PT.func.isRequired,
   resetBehandlingerState: PT.func.isRequired,
@@ -370,13 +274,8 @@ Saksbehandling.propTypes = {
   resetMedlemskapsperiodeState: PT.func.isRequired,
   startOgVisOppfriskModal: PT.func.isRequired,
   skjulMenypanel: PT.func.isRequired,
-  tilbakeleggOppgave: PT.func.isRequired,
   tilForsiden: PT.func.isRequired,
-  visAvsluttSakSomBortfaltDialogHandle: PT.func.isRequired,
-  visAvslagSoknadDialogHandle: PT.func.isRequired,
-  visHenleggDialogHandle: PT.func.isRequired,
   visOppfriskModal: PT.func.isRequired,
-  visRevurderFagsakDialogHandle: PT.func.isRequired,
 };
 
 Saksbehandling.defaultProps = {
@@ -399,15 +298,12 @@ const mapStateToProps = (state) => ({
     behandlingsgrunnlagSelectors.MottaksdatoSelector(state)
   ),
   behandlingsresultat: behandlingsresultatSelectors.BehandlingsresultatSelector(state),
-  behandlingstema: behandlingerSelectors.BehandlingstemaKodeSelector(state),
-  behandlingsstatus: behandlingerSelectors.BehandlingsstatusKodeSelector(state),
   dokumenter: dokumenterSelectors.AlleFysiskeDokumentSelector(state),
   dokumentOversikt: dokumenterSelectors.DokumentOversiktSelector(state),
   fagsak: fagsakSelectors.FagsakSelector(state),
   fagsakStatusKode: fagsakSelectors.FagsakStatusSelector(state),
   landkoder: landkoderSelectors.LandkoderFraSakstypeSelector(state),
   oppsummering: behandlingerSelectors.OppsummeringSelector(state),
-  person: behandlingerSelectors.PersonSelector(state),
   redigerbart: redigerbartSelectors.RedigerbartSelector(state),
   skjema: formSelectors.SoknadenFormSelector(state).values,
   soknadForm: formSelectors.SoknadenFormSelector(state),
