+++ src/sider/trygdeavtale/saksbehandling.js
@@ -12,10 +12,7 @@ import SideDialog from "../../felleskomponenter/sideDialog/sideDialog";
 import { AvslaattSoknad, HenlagtSak } from "../eu_eøs/saksbehandling/komponenter/stegErstatter";
 import SideOppsummering from "../../felleskomponenter/oppsummering/sideOppsummering";
 import SaksoversiktLenke from "../../felleskomponenter/saksoversiktLenke";
-import Behandlingsstatus from "../../felleskomponenter/behandlingsstatus";
 import { SoknadMenypanelForm } from "../../felleskomponenter/menypanelForm";
-import { FeatureToggle } from "../../featuretoggle";
-import Legacybehandlingsmeny from "./legacybehandlingsmeny";
 
 import { behandlingsgrunnlagOperations, behandlingsgrunnlagSelectors } from "../../ducks/behandlingsgrunnlag";
 import { behandlingsresultatOperations, behandlingsresultatSelectors } from "../../ducks/behandlingsresultat";
@@ -31,61 +28,8 @@ import { formSelectors } from "../../ducks/form";
 import Stegvelger from "./stegvelger";
 import "./saksbehandling.css";
 
-const behandlingsstatusMap = {
-  [MKV.Koder.behandlinger.behandlingsstatus.VURDER_DOKUMENT]: [
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-      term: MKV.Terms.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-    },
-  ],
-  [MKV.Koder.behandlinger.behandlingsstatus.SVAR_ANMODNING_MOTTATT]: [
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-      term: MKV.Terms.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-    },
-  ],
-  [MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_UTL]: [
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_PART,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-      term: MKV.Terms.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-    },
-  ],
-  [MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_PART]: [
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-      term: MKV.Terms.behandlinger.behandlingsstatus.AVVENT_DOK_UTL,
-    },
-    {
-      kode: MKV.Koder.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-      term: MKV.Terms.behandlinger.behandlingsstatus.UNDER_BEHANDLING,
-    },
-  ],
-  [MKV.Koder.behandlinger.behandlingsstatus.UNDER_BEHANDLING]: [],
-};
-
 const Saksbehandling = ({
   annenBehandlingOppfriskes,
-  apneTidligereBehandlinger,
   arbeidsland,
   behandlingOppfriskes,
   behandlingsgrunnlag,
@@ -93,8 +37,6 @@ const Saksbehandling = ({
   behandlingsgrunnlagPeriodeFom,
   behandlingsgrunnlagPeriodeTom,
   behandlingsresultat,
-  behandlingstema,
-  behandlingsstatus,
   dokumenter,
   dokumentOversikt,
   fagsak,
@@ -106,12 +48,10 @@ const Saksbehandling = ({
   hentLandkoder,
   hentLovvalgsperiode,
   hentFagsaker,
-  lagreOgLukk,
   location,
   match,
   oppfriskOgLastInnSaksopplysninger,
   oppsummering,
-  person,
   redigerbart,
   resetBehandlingerState,
   resetBehandlingsgrunnlagState,
@@ -120,12 +60,7 @@ const Saksbehandling = ({
   startOgVisOppfriskModal,
   soknadForm,
   tilForsiden,
-  tilbakeleggOppgave,
-  visAvslagSoknadDialogHandle,
-  visAvsluttSakSomBortfaltDialogHandle,
-  visHenleggDialogHandle,
   visOppfriskModal,
-  visRevurderFagsakDialogHandle,
   lovvalgsperiodeTom,
 }) => {
   const [behandlingID, setBehandlingID] = useState(-1);
@@ -208,9 +143,7 @@ const Saksbehandling = ({
   const visStegVelger = !erHenlagtSak && !erAvslaattSoknad && behandlingsgrunnlagErKlart && !behandlingIDHarEndretSeg;
   return (
     <>
-      <FeatureToggle togglename="melosys.design.PERSONLINJE">
-        {(status) => status === "enabled" && <Personlinje />}
-      </FeatureToggle>
+      <Personlinje />
       <div id="main-container" className="main-container">
         <div className="saksbehandling">
           <Nav.Container fluid>
@@ -230,39 +163,14 @@ const Saksbehandling = ({
               </Nav.Column>
               <Nav.Column xs="5">
                 <SideOppsummering
-                  behandlingstema={behandlingstema}
-                  redigerbart={redigerbart}
-                  fagsak={fagsak}
                   oppsummering={oppsummering}
-                  person={person}
+                  fagsak={fagsak}
                   arbeidsland={arbeidsland}
+                  lovvalgsperiodeFom={behandlingsgrunnlagPeriodeFom}
+                  lovvalgsperiodeTom={lovvalgsperiodeTom || behandlingsgrunnlagPeriodeTom}
                   behandlingsgrunnlagPeriodeFom={behandlingsgrunnlagPeriodeFom}
                   behandlingsgrunnlagPeriodeTom={behandlingsgrunnlagPeriodeTom}
                   behandlingsgrunnlagMottaksdato={behandlingsgrunnlagMottaksdato}
-                  lovvalgsperiodeFom={behandlingsgrunnlagPeriodeFom}
-                  lovvalgsperiodeTom={lovvalgsperiodeTom || behandlingsgrunnlagPeriodeTom}
-                  renderBehandlingsmeny={() => (
-                    <Legacybehandlingsmeny
-                      redigerbart={redigerbart}
-                      lagreOgLukkHandle={lagreOgLukk}
-                      tilbakeleggeHandle={tilbakeleggOppgave}
-                      oppfriskSaksopplysningerHandle={visOppfriskModal}
-                      visHenleggDialogHandle={visHenleggDialogHandle}
-                      visAvsluttSakSomBortfaltDialogHandle={visAvsluttSakSomBortfaltDialogHandle}
-                      visAvslagSoknadDialogHandle={visAvslagSoknadDialogHandle}
-                      apneTidligereBehandlinger={apneTidligereBehandlinger}
-                      visRevurderFagsakDialogHandle={visRevurderFagsakDialogHandle}
-                      behandlingsstatus={behandlingsstatus}
-                    />
-                  )}
-                  renderBehandlingsstatus={() => (
-                    <Behandlingsstatus
-                      behandlingID={behandlingID}
-                      redigerbart={redigerbart}
-                      oppsummering={oppsummering}
-                      behandlingsstatusMap={behandlingsstatusMap}
-                    />
-                  )}
                 />
                 <SaksoversiktLenke />
                 <SideDialog
@@ -290,8 +198,6 @@ Saksbehandling.propTypes = {
   behandlingsgrunnlagPeriodeTom: PT.string.isRequired,
   behandlingsgrunnlagMottaksdato: PT.string.isRequired,
   behandlingsresultat: MPT.Behandlingsresultat.isRequired,
-  behandlingstema: PT.string.isRequired,
-  behandlingsstatus: PT.string.isRequired,
   dokumenter: PT.array.isRequired,
   dokumentOversikt: PT.array.isRequired,
   fagsak: MPT.Fagsak,
@@ -299,11 +205,9 @@ Saksbehandling.propTypes = {
   location: PT.object.isRequired,
   match: PT.object.isRequired,
   oppsummering: MPT.Behandlinger.Oppsummering,
-  person: MPT.Behandlinger.Saksopplysninger.Person.isRequired,
   redigerbart: PT.bool,
   soknadForm: PT.object.isRequired,
   // Funcs
-  apneTidligereBehandlinger: PT.func.isRequired,
   hentBehandling: PT.func.isRequired,
   hentBehandlingsgrunnlag: PT.func.isRequired,
   hentBehandlingsresultat: PT.func.isRequired,
@@ -311,20 +215,14 @@ Saksbehandling.propTypes = {
   hentLandkoder: PT.func.isRequired,
   hentLovvalgsperiode: PT.func.isRequired,
   hentFagsaker: PT.func.isRequired,
-  lagreOgLukk: PT.func.isRequired,
   oppfriskOgLastInnSaksopplysninger: PT.func.isRequired,
-  tilbakeleggOppgave: PT.func.isRequired,
   tilForsiden: PT.func.isRequired,
   resetBehandlingerState: PT.func.isRequired,
   resetBehandlingsgrunnlagState: PT.func.isRequired,
   resetFagsakState: PT.func.isRequired,
   skjulMenypanel: PT.func.isRequired,
   startOgVisOppfriskModal: PT.func.isRequired,
-  visAvsluttSakSomBortfaltDialogHandle: PT.func.isRequired,
-  visAvslagSoknadDialogHandle: PT.func.isRequired,
-  visHenleggDialogHandle: PT.func.isRequired,
   visOppfriskModal: PT.func.isRequired,
-  visRevurderFagsakDialogHandle: PT.func.isRequired,
   lovvalgsperiodeTom: PT.string.isRequired,
 };
 
@@ -348,14 +246,11 @@ const mapStateToProps = (state) => ({
     behandlingsgrunnlagSelectors.MottaksdatoSelector(state)
   ),
   behandlingsresultat: behandlingsresultatSelectors.BehandlingsresultatSelector(state),
-  behandlingstema: behandlingerSelectors.BehandlingstemaKodeSelector(state),
-  behandlingsstatus: behandlingerSelectors.BehandlingsstatusKodeSelector(state),
   dokumenter: dokumenterSelectors.AlleFysiskeDokumentSelector(state),
   dokumentOversikt: dokumenterSelectors.DokumentOversiktSelector(state),
   fagsak: fagsakSelectors.FagsakSelector(state),
   fagsakStatusKode: fagsakSelectors.FagsakStatusSelector(state),
   oppsummering: behandlingerSelectors.OppsummeringSelector(state),
-  person: behandlingerSelectors.PersonSelector(state),
   redigerbart: redigerbartSelectors.RedigerbartSelector(state),
   soknadForm: formSelectors.SoknadenFormSelector(state),
   lovvalgsperiodeTom: Utils.dato.formatterDatoTilNorsk(lovvalgsperioderSelectors.TomDatoSelector(state)),
